Don't burn active probing-resistant PTs.
The GFW used to block bridges at an IP:port level, but a while ago, it
started to block bridges on an IP address level instead.  This means
that if a bridge runs obfs4 (which is active probing-resistant) and
obfs3 (which is not active probing-resistant), and happens to hand out
obfs3, the GFW would manage to probe the bridge, and block the entire IP
address, *including* obfs4.

Our solution is to not hand out probe-able protocols if a bridge runs an
active probing-resistant protocol such as obfs4 or scramblesuit.

This fixes bug 28655: <https://bugs.torproject.org/28655>